python_tools/requirements.txt
paddleocr
unittest
pyperclip
pillow